% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{aggregate_wl}
\alias{aggregate_wl}
\title{Aggregate wavelengths}
\usage{
aggregate_wl(x, FUN = "mean", ...)
}
\arguments{
\item{x}{a hyperSpec object.}

\item{FUN}{function to be used for aggregation.}

\item{...}{(named) selection of bands or band ranges.}
}
\value{
a hyperspec object with aggregated spc
}
\description{
select certain bands or aggregate wavelength ranges of a hyperSpec object.
}
\examples{
\dontrun{
aggregate_wl(chondro, "mean", band_a = 938, band_b = 1230 ~ 1260)
}
}
