% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/als.R
\name{als}
\alias{als}
\title{Alternating least squares multivariate curve resolution (MCR-ALS)}
\usage{
als(x, ncomp = NULL, S_init = NULL, prefix = "component")
}
\arguments{
\item{x}{a hyperSpec object.}

\item{ncomp}{integer; number of pure components, does not need to be specified
if S_init is provided.}

\item{S_init}{a hyperSpec object; initial estimates of pure spectral components,
if S_init is NULL, initial components are estimated
using \code{\link{opa}}.}

\item{prefix}{character; a prefix to name the pure spectra.}
}
\value{
a list with the following components:
  \describe{
    \item{data}{the inpout hyperSpec object with an additional "component" colunb in @data}
    \item{component_spectra}{a hyperSpec object containing the component spectra}
    \item{summary_stats}{a named vector containing lof, rms and r2}
  }
}
\description{
This is a wrapper around a modified version of \link[alsace]{doALS}
  from the package {alsace} to be used with objects of class hyperSpec.
}
\seealso{
\link[alsace]{doALS}, \link[ALS]{als}
}
