% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{diff,hyperSpec-method}
\alias{diff,hyperSpec-method}
\alias{diff}
\title{differences on hyperSpec spectra (S4 method)}
\usage{
\S4method{diff}{hyperSpec}(x, lag = 1, differences = 1)
}
\arguments{
\item{x}{a numeric vector or matrix containing the values to be
    differenced.}

\item{lag}{an integer indicating which lag to use.}

\item{differences}{an integer indicating the order of the difference.}
}
\value{
a hyperSpec object.
}
\description{
S4 method for the generic function \code{diff()} and the hyperSpec class.
}
\details{
All columns of \code{x@data$spc} are kept and spectra padded with NAs accordingly.
}
\examples{
\dontrun{
plotspc(diff(chondro, 1, 1))
}
}
\seealso{
\code{\link[base]{diff}}
}
