% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperSpec.utils-package.R
\docType{package}
\name{hyperSpec.utils-package}
\alias{hyperSpec.utils}
\alias{hyperSpec.utils-package}
\title{hyperSpec.utils: Utility functions to work with objects of the hyperSpec class}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/konradmayer/hyperSpec.utils}
  \item Report bugs at \url{https://github.com/konradmayer/hyperSpec.utils/issues}
}

}
\author{
\strong{Maintainer}: Konrad Mayer \email{konrad.mayer@boku.ac.at} (\href{https://orcid.org/0000-0002-8803-068X}{ORCID})

}
\keyword{internal}
