% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ica.R
\name{ica}
\alias{ica}
\title{Spectral unmixing: Independent Component Analysis (ICA)}
\usage{
ica(x, ncomp, prefix = "basis", ...)
}
\arguments{
\item{x}{a hyperSpec object.}

\item{ncomp}{integer; number of pure components.}

\item{prefix}{character; a prefix to name the pure spectra.}

\item{...}{further arguments to be passed down to \code{\link[fastICA]{fastICA}}.}
}
\value{
a list with the following components:
  \describe{
    \item{coefficients}{coefficient matrix}
    \item{basis}{a hyperSpec object containing the basis (component) spectra}
    \item{fit}{the ica fit as returned by  \code{\link[fastICA]{fastICA}}}
  }
}
\description{
This is a wrapper around \code{\link[fastICA]{fastICA}}
  from the package {fastICA} to be used with objects of class hyperSpec.
}
\references{
A. Hyvarinen, "Fast and robust fixed-point algorithms for independent component analysis," in IEEE Transactions on Neural Networks, vol. 10, no. 3, pp. 626-634, May 1999 \url{https://doi.org/10.1109/72.761722}
}
\seealso{
\code{\link[fastICA]{fastICA}}
}
