% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.R
\name{integrate_rectangular}
\alias{integrate_rectangular}
\title{Rectangular (midpoint) integration}
\usage{
integrate_rectangular(x, y, correct_edges = TRUE)
}
\arguments{
\item{x}{numeric.}

\item{y}{numeric.}

\item{correct_edges}{logical; if set TRUE, the widths of the segments at the
range edges (the first and last point) are corrected to only incorporate
the half width to the right, resp. to the left of the point.}
}
\value{
numeric; the approximated integral.
}
\description{
Rectangular (midpoint) integration
}
