% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.R
\name{integrate_trapezoid}
\alias{integrate_trapezoid}
\title{Trapezoid integration}
\usage{
integrate_trapezoid(x, y)
}
\arguments{
\item{x}{numeric.}

\item{y}{numeric.}
}
\value{
numeric; the approximated integral.
}
\description{
Trapezoid integration
}
