% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.R
\name{integrate_wl}
\alias{integrate_wl}
\title{Integrate over wavelengths}
\usage{
integrate_wl(x, ..., FUN = integrate_trapezoid)
}
\arguments{
\item{x}{a hyperSpec object.}

\item{...}{(named) selection of wavelength ranges (specify as in subsetting a
hyperSpec object by using a tilde).}

\item{FUN}{function to be used for integration (trapezoid approximation as
default).}
}
\value{
a hyperspec object with integrated spc
}
\description{
Integrate over named wavelength ranges of a hyperSpec object.
}
\examples{
\dontrun{
integrate_wl(chondro, integral_a = 930 ~ 940, integral_b = 1230 ~ 1260)
}
}
