% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation.R
\name{interpolation}
\alias{interpolation}
\alias{spc_linear_interpolation}
\alias{spc_spline_interpolation}
\title{Interpolation of spectra onto a new wavelength axis}
\usage{
spc_linear_interpolation(spc, newx = hyperSpec::wl(spc), ...)

spc_spline_interpolation(spc, newx = hyperSpec::wl(spc), ...)
}
\arguments{
\item{spc}{an object of class \code{hyperSpec}}

\item{newx}{numeric; a new wavelength (or wavenumber) vector}

\item{...}{additional arguments to be handed further to \link[hyperSpec]{spc.smooth.spline} or \link[stats]{approx}.}
}
\description{
Interpolation of spectra onto a new wavelength axis
}
