% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_hyperSpec.R
\name{is_hyperSpec}
\alias{is_hyperSpec}
\alias{is.hyperSpec}
\alias{is.hyperSpecMap}
\alias{is_hyperSpecMap}
\title{Test if the object is of class hyperSpec}
\usage{
is_hyperSpec(x)

is_hyperSpecMap(x)
}
\arguments{
\item{x}{An object}
}
\description{
This function returns TRUE for objects of class \code{hyperSpec}, and FALSE for all other objects.
The variant \code{is.hyperSpecMap} returns only TRUE if additionally variables "x" and "y" are present in \code{@data}.
}
