% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_wl_equidistant}
\alias{is_wl_equidistant}
\title{Test if the wavelength vector of a hyperSpec object is equally spaced}
\usage{
is_wl_equidistant(x)
}
\arguments{
\item{x}{a hyperSpec object}
}
\value{
logical
}
\description{
Test if the wavelength vector of a hyperSpec object is equally spaced
}
