% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_wl_ordered}
\alias{is_wl_ordered}
\title{Test if the wavelength vector of a hyperSpec object is in order (ascending)}
\usage{
is_wl_ordered(x)
}
\arguments{
\item{x}{a hyperSpec object}
}
\value{
logical
}
\description{
Test if the wavelength vector of a hyperSpec object is in order (ascending)
}
