% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{normalization}
\alias{normalization}
\alias{minmax_normalization}
\alias{snv_normalization}
\alias{vector_normalization}
\alias{area_normalization}
\alias{band_normalization}
\title{Normalization}
\usage{
minmax_normalization(x)

snv_normalization(x)

vector_normalization(x)

area_normalization(x, FUN = c("mean", "sum"))

band_normalization(x, band)
}
\arguments{
\item{x}{A \code{hyperSpec} object}

\item{FUN}{character; "mean" or "sum" to select the function to use for the area normalization.  Using "sum", the integral under the spectrum gets 1, using "mean" the spectrum scales with intensities around 1.}

\item{band}{numeric or formula; specifying the band or wavelength range to use for band normalization.}
}
\value{
A \code{hyperSpec} object with normalized intensities
}
\description{
Perform normalization on objects of class \code{hyperSpec}.
}
