% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omp.R
\name{omp}
\alias{omp}
\title{Spectral mixture analysis: Orthogonal Matching Pursuit (OMP)}
\usage{
omp(
  targets,
  references,
  tol = 0.2,
  references_ids = NULL,
  parallel = FALSE,
  ncores = 4
)
}
\arguments{
\item{targets}{a hyperSpec object.}

\item{references}{a hyperSpec object containing the spectra to be combined in
in a linear manner.}

\item{tol}{The tolerance value to terminate the algorithm, see
\code{\link[Rfast]{ompr}}. As the function is run with
\code{method = "SSE"}, the algorithm stops if the relative change in the
sum of squared errors is equal to or less than \code{tol}.}

\item{references_ids}{either a single string or numeric specifying the column with names
for the references in the object `references`, or a character vector with
a name per reference spectrum.}

\item{parallel}{logical; If this is \code{TRUE},
\code{\link[parallel]{parApply}} is used. Please be aware that the function
handles starting and stoping of the cluster (type = "PSOCK").}

\item{ncores}{number of logical cores to use for parallel processing.
Be aware that more is not necessarily better, as starting new processes
adds overhead which potentially is bigger than the time savings from
parallel processing.}
}
\value{
a list containing the following members:
  \describe{
    \item{coefficients}{coefficient matrix}
    \item{basis}{the references object as supplied}
    \item{fit}{the return value of the multivariate call to `stats::lm()` or a list of return values of the outputs of `nnls::nnls()` for each target spectrum.}
  }
}
\description{
Use the Orthogonal Matching Pursuit to estimate sparse
  coefficients for a linear combination of reference spectra to represent
  one or multiple target spectra.
}
\details{
This is a wrapper around \code{\link[Rfast]{ompr}}. The function is
  run with \code{method = "SSE"}.
}
\references{
Pati, Y. C., R. Rezaiifar, Y. C. Pati R. Rezaiifar, and P. S. Krishnaprasad. “Orthogonal Matching Pursuit: Recursive Function Approximation with Applications to Wavelet Decomposition.” In Proceedings of the 27 Th Annual Asilomar Conference on Signals, Systems, and Computers, 40–44, 1993.
}
