% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmap.R
\name{plotmap_viridis}
\alias{plotmap_viridis}
\title{Plotmap using viridis}
\usage{
plotmap_viridis(
  object,
  model = spc ~ x * y,
  func = mean,
  func.args = list(),
  ...
)
}
\arguments{
\item{object}{the \code{hyperSpec} object}

\item{model}{formula specifying the columns of object that are to be
displayed by \code{\link[lattice]{levelplot}}}

\item{func}{Before plotting, \code{plotmap} applies function
  \code{func} with the arguments given in the list \code{func.args} to each
  of the spectra. Thus a single summary value is displayed for each of the
  spectra.

This can be suppressed manually by setting \code{func} to NULL. It is automatically suppressed if
\code{.wavelength} appears in the formula.}

\item{func.args}{Before plotting, \code{plotmap} applies function
  \code{func} with the arguments given in the list \code{func.args} to each
  of the spectra. Thus a single summary value is displayed for each of the
  spectra.

This can be suppressed manually by setting \code{func} to NULL. It is automatically suppressed if
\code{.wavelength} appears in the formula.}

\item{...}{further arguments are passed down the call chain, and finally
to \code{\link[lattice]{levelplot}}}
}
\description{
Use the viridis colorramp in plotmap as default. The model can
  contain the special column name .wavelength to specify the wavelength axis.
}
