% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plotspc_rev}
\alias{plotspc_rev}
\title{Plotting Spectra}
\usage{
plotspc_rev(object, ...)
}
\arguments{
\item{object}{the \code{hyperSpec} object}

\item{...}{additional arguments handed to \code{\link[hyperSpec]{plotspc}}.}
}
\value{
\code{plotspc} invisibly returns a list with

\item{x}{the abscissa coordinates of the plotted spectral data points}

\item{y}{a matrix the ordinate coordinates of the plotted spectral data
  points}

\item{wavelengths}{the wavelengths of the plotted spectral data points}

This can be used together with \code{\link[hyperSpec]{spc.identify}}.

a list containing \item{offsets}{numeric with the yoffset for each group in
\code{stacked}} \item{groups}{numeric with the group number for each spectrum} \item{levels}{if
\code{stacked} is a factor, the levels of the groups}
}
\description{
Simple wrapper for \code{\link[hyperSpec]{plotspc}} to execute
  the function with reversed x-axis as default.
}
