% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_opus.R
\name{read_opus_single}
\alias{read_opus_single}
\title{Import a single Bruker OPUS binary file}
\usage{
read_opus_single(file_path, scale_y = TRUE, read_info = FALSE)
}
\arguments{
\item{file_path}{character; path to the Bruker OPUS binary file.}

\item{scale_y}{logical; use the y scaling factor provided within the
binary file.}

\item{read_info}{logical; read info stored in the text information data
block.}
}
\value{
a hyperSpec object.
}
\description{
Function to import single OPUS binary files; supports
  reading meta data stored in the text information data block. At the current
  state, no other information (e.g. instrumental parameters) such as INS,
  LWN, HUM, SRC, BMS, ZFF, DAT, TIM are processed. Consider this function as
  experimental.
}
