% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocorrelation.R
\name{spatial_ac}
\alias{spatial_ac}
\title{Spatial autcorrelation}
\usage{
spatial_ac(x, neigh = 2, colname = "spatial_ac", ...)
}
\arguments{
\item{x}{a hyperSpec object.}

\item{neigh}{neighborhood size as defined in \link[ncf]{lisa.nc}.}

\item{colname}{name of the data column to be added to \code{x@data},
containing the mean correlation within the neighborhood.}

\item{...}{additional arguments handed further to \link[ncf]{lisa.nc}.}
}
\value{
a hyperSpec object such as given for \code{x} with an additional
  column in \code{x@data}.
}
\description{
This function is a wrapper around \link[ncf]{lisa.nc} to
  calculate spatial autocorrelation of a spectral map.
}
