% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{spcmap_dim}
\alias{spcmap_dim}
\title{Dimensions of a hyperSpec map}
\usage{
spcmap_dim(x)
}
\arguments{
\item{x}{hyperSpec object with variables "x" and "y" present in \code{@data}.}
}
\value{
a named vector with the number of x and y positions as well as the number of descrete wavelengths ("nwl")
}
\description{
Dimensions of a hyperSpec map
}
\examples{
library(hyperSpec)
spcmap_dim(chondro)
}
