% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spcmap_explorer.R
\name{spcmap_explorer}
\alias{spcmap_explorer}
\title{Interactive spectral map explorer}
\usage{
spcmap_explorer(
  hyperspec_obj,
  fixed_y = TRUE,
  flip = FALSE,
  startband = 1600,
  metavar = NULL
)
}
\arguments{
\item{hyperspec_obj}{A \code{hyperSpec} object with variables "x" and "y" in \code{@data}.}

\item{fixed_y}{logical or numeric; if TRUE the initial y axis range is set to the intensity range of the complete dataset until the axis is manually modified or autoscale activated; alternatively provide a numeric vector of length 2 to set a fixed y range; if FALSE autoscale is activated.}

\item{flip}{logical; if TRUE the x and y axes are transposed.}

\item{startband}{numeric; band for the displayed image at startup.}

\item{metavar}{character; a variable in \code{@data} can be selected as an alternative to the intensity map (e.g. to display cluster results).}
}
\description{
Shiny app to interactively explore spatially resolved \code{hyperSpec} objects.
}
