% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocorrelation.R
\name{spectral_ac}
\alias{spectral_ac}
\title{Spectral autocorrelation}
\usage{
spectral_ac(x, lag = 1, colname = "spectral_ac")
}
\arguments{
\item{x}{a hyperSpec object.}

\item{lag}{lag at which the autocorrelation is calculated.}

\item{colname}{name of the data column to be added to \code{x@data},
containing the autocorrelation coefficient.}
}
\value{
a hyperSpec object such as given for \code{x} with an additional
  column in \code{x@data}.
}
\description{
Autocorrelation of a spectrum. Generally, a high autocorrelation
  is associated with high signal, while white noise is not autocorrelated.
}
