% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_wl_equidistant}
\alias{is_wl_equidistant}
\title{Test if the wavelength vector of a hyperSpec object is equally spaced}
\usage{
is_wl_equidistant(x, tolerance = sqrt(.Machine$double.eps))
}
\arguments{
\item{x}{a hyperSpec object}

\item{tolerance}{maximum difference to accept (to account for machine precision)}
}
\value{
logical
}
\description{
Test if the wavelength vector of a hyperSpec object is equally spaced
}
