% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_combination.R
\name{linear_combination}
\alias{linear_combination}
\title{Linear combination of spectra}
\usage{
linear_combination(
  targets,
  references,
  references_ids = NULL,
  nonnegative = FALSE
)
}
\arguments{
\item{targets}{a hyperSpec object.}

\item{references}{a hyperSpec object containing the spectra to be combined in
in a linear manner.}

\item{references_ids}{either a single string specifying the column with names
for the references in the object `references`, or a character vector with
a name per reference spectrum.}

\item{nonnegative}{logical; set TRUE to restrict coefficients to be positive.}
}
\value{
a list containing the following members:
  \describe{
    \item{data}{the target hyperSpec object with additional columns in `@data` for the coefficients for each of the references as well as summary statistics (RMSE and R2)}
    \item{references}{the references object as supplied}
    \item{fit}{the return value of the multivariate call to `stats::lm()` or a list of return values of the outputs of `nnls::nnls()` for each target spectrum.}
  }
}
\description{
Use linear regression resp. non negative
  linear regression (package {nnls}) without intercept to estimate
  coefficients for a linear combination of reference spectra to represent
  one or multiple target spectra.
}
