% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf.R
\name{nmf}
\alias{nmf}
\title{Spectral unmixing: Nonnegative matrix factorization (NMF)}
\usage{
nmf(
  x,
  ncomp = NULL,
  nrun = 1,
  method = "brunet",
  seed = "ica",
  prefix = "component",
  ...
)
}
\arguments{
\item{x}{a hyperSpec object.}

\item{ncomp}{integer; number of pure components (factorization rank).}

\item{nrun}{number of runs}

\item{method}{character; use `NMF::getNMFMethod()` to get valid Methods and
refer to the documentation of `NMF::nmf()` for further information.}

\item{seed}{character; one of `NMF::getNMFSeed()`. Refer to the documentation
of `NMF::nmf()` for further information.}

\item{prefix}{character; a prefix to name the pure spectra.}

\item{...}{further arguments to be passed down to \link[NMF]{nmf}.}
}
\value{
a list with the following components:
  \describe{
    \item{data}{the input hyperSpec object with additional "component" columns in @data, holding the coefficients}
    \item{component_spectra}{a hyperSpec object containing the component spectra}
    \item{fit}{the nmf fit as returned by `NMF::nmf()`}
  }
}
\description{
This is a wrapper around \link[NMF]{nmf}
  from the package {NMF} to be used with objects of class hyperSpec.
}
\seealso{
\link[NMF]{nmf}
}
