% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_opus.R
\name{read_opus}
\alias{read_opus}
\alias{read.opus}
\title{Read Bruker OPUS binary files}
\usage{
read_opus(file_paths, scale_y = TRUE, read_info = FALSE, wl = 1)
}
\arguments{
\item{file_paths}{a character vector of file paths to Bruker OPUS binary
files}

\item{scale_y}{logical; use the y scaling factor provided within the
binary file.}

\item{read_info}{logical; read info stored in the text information data
block.}

\item{wl}{numeric; either of length one giving the index of the spectrum to
use the associated wavelength vector, a numeric vector of length >1
specifying a new wavelength vector, \code{NULL} to use a merged wavelength
vector or "interpolate" to interpolate missing values after using the
merged vector.}
}
\value{
a hyperSpec object.
}
\description{
This function is a wrapper around \code{\link{read_opus_single}}
  to import multiple OPUS binary files; supports
  reading meta data stored in the text information data block. At the current
  state, no other information (e.g. instrumental parameters) such as INS,
  LWN, HUM, SRC, BMS, ZFF, DAT, TIM are processed. Consider this function as
  experimental and sanity check imported spectra.
}
\details{
When files to import don't share common wavelength vectors, a
  collapsed hyperSpec object (having a wavelength vector merged from all the
  individual vectors) may become very large with a lot of missing
  values. \code{read_opus()} can interpolate the individual spectra onto a
  common wavelength vector prior collapsing, either by choosing the
  wavelength vector of one of the imported files (giving a single integer) or
  by providing a new wavelength vector. Alternatively, if \code{wl} is
  \code{NULL}, the merged wavelength vector is used, potentially resulting in
  a sparse spectral matrix. If set to "interpolate" the missing values are
  interpolated after using the merged vector.
}
