% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_opus.R
\name{read_opus}
\alias{read_opus}
\alias{read.opus}
\title{Read Bruker OPUS binary files}
\usage{
read_opus(file_paths, scale_y = TRUE, read_info = FALSE, interpolate = FALSE)
}
\arguments{
\item{file_paths}{a character vector of file paths to Bruker OPUS binary
files}

\item{scale_y}{logical; use the y scaling factor provided within the
binary file.}

\item{read_info}{logical; read info stored in the text information data
block.}

\item{interpolate}{logical, unify the wavelength axis (interpolate at
not shared wavelengths) among the individual imported spectra using
\code{\link[hyperSpec]{spc.NA.approx}}. Not necessary if all spectra share
the same wavelength axis.}
}
\value{
a hyperSpec object.
}
\description{
This function is a wrapper around \code{\link{read_opus_single}}
  to import multiple OPUS binary files; supports
  reading meta data stored in the text information data block. At the current
  state, no other information (e.g. instrumental parameters) such as INS,
  LWN, HUM, SRC, BMS, ZFF, DAT, TIM are processed. Consider this function as
  experimental.
}
