% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spcmap2array.R
\name{spcmap2array}
\alias{spcmap2array}
\title{hyperSpec map to array}
\usage{
spcmap2array(x)
}
\arguments{
\item{x}{A \code{hyperSpec} object with variables "x" and "y" in \code{@data}.}
}
\value{
An array with the dimensions "x" "y" and "nwl", with the wavelength vector attached as an attribute.
}
\description{
Convert spatially resolved object of class hyperSpec to an array.
}
